Move backoffice-only APIs to new package
