+++ src/main/java/pt/ua/deti/tqs/backend/controllers/backoffice/CityController.java
@@ -1,4 +1,4 @@
-package pt.ua.deti.tqs.backend.controllers;
+package pt.ua.deti.tqs.backend.controllers.backoffice;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -14,7 +14,7 @@ import pt.ua.deti.tqs.backend.services.CityService;
 import java.util.List;
 
 @RestController
-@RequestMapping("api/city")
+@RequestMapping("backoffice/city")
 @Tag(name = "City")
 @AllArgsConstructor
 public class CityController {
