+++ src/test/java/pt/ua/deti/tqs/backend/controllers/StatsControllerTest.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
+import pt.ua.deti.tqs.backend.controllers.backoffice.StatsController;
 import pt.ua.deti.tqs.backend.services.StatsService;
 
 import static org.hamcrest.Matchers.is;
